# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multiple Disease Predictor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PredictionResult {
  /** 0 = negative, 1 = positive */
  prediction: number;
  /** Human-readable result label */
  label: string;
  /** Confidence score between 0 and 1 */
  confidence?: number;
  /** Descriptive result message */
  message: string;
}

export interface ErrorResponse {
  error: string;
  details?: string;
}

export interface DiabetesInput {
  /** Number of times pregnant */
  pregnancies: number;
  /** Plasma glucose concentration */
  glucose: number;
  /** Diastolic blood pressure (mm Hg) */
  bloodPressure: number;
  /** Triceps skin fold thickness (mm) */
  skinThickness: number;
  /** 2-Hour serum insulin (mu U/ml) */
  insulin: number;
  /** Body mass index */
  bmi: number;
  /** Diabetes pedigree function */
  diabetesPedigreeFunction: number;
  /** Age (years) */
  age: number;
}

export interface HeartInput {
  /** Age in years */
  age: number;
  /** Sex (1 = male, 0 = female) */
  sex: number;
  /** Chest pain type (0-3) */
  cp: number;
  /** Resting blood pressure (mm Hg) */
  trestbps: number;
  /** Serum cholestoral in mg/dl */
  chol: number;
  /** Fasting blood sugar > 120 mg/dl (1 = true, 0 = false) */
  fbs: number;
  /** Resting electrocardiographic results (0-2) */
  restecg: number;
  /** Maximum heart rate achieved */
  thalach: number;
  /** Exercise induced angina (1 = yes, 0 = no) */
  exang: number;
  /** ST depression induced by exercise */
  oldpeak: number;
  /** Slope of peak exercise ST segment (0-2) */
  slope: number;
  /** Number of major vessels colored by fluoroscopy (0-3) */
  ca: number;
  /** Thalassemia (0 = normal, 1 = fixed defect, 2 = reversable defect) */
  thal: number;
}

export interface ParkinsonsInput {
  /** MDVP Fo(Hz) - Average vocal fundamental frequency */
  mdvpFo: number;
  /** MDVP Fhi(Hz) - Maximum vocal fundamental frequency */
  mdvpFhi: number;
  /** MDVP Flo(Hz) - Minimum vocal fundamental frequency */
  mdvpFlo: number;
  /** MDVP Jitter(%) */
  mdvpJitter: number;
  /** MDVP Jitter(Abs) */
  mdvpJitterAbs: number;
  /** MDVP RAP */
  mdvpRap: number;
  /** MDVP PPQ */
  mdvpPpq: number;
  /** Jitter DDP */
  jitterDdp: number;
  /** MDVP Shimmer */
  mdvpShimmer: number;
  /** MDVP Shimmer(dB) */
  mdvpShimmerDb: number;
  /** Shimmer APQ3 */
  shimmerApq3: number;
  /** Shimmer APQ5 */
  shimmerApq5: number;
  /** MDVP APQ */
  mdvpApq: number;
  /** Shimmer DDA */
  shimmerDda: number;
  /** NHR - Noise-to-Harmonics Ratio */
  nhr: number;
  /** HNR - Harmonics-to-Noise Ratio */
  hnr: number;
  /** RPDE - Recurrence Period Density Entropy */
  rpde: number;
  /** DFA - Detrended Fluctuation Analysis */
  dfa: number;
  /** Spread1 - nonlinear measure */
  spread1: number;
  /** Spread2 - nonlinear measure */
  spread2: number;
  /** D2 - nonlinear dynamical complexity */
  d2: number;
  /** PPE - Pitch Period Entropy */
  ppe: number;
}