# lib/api-zod/src/generated/types/diabetesInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multiple Disease Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface DiabetesInput {
  /** Number of times pregnant */
  pregnancies: number;
  /** Plasma glucose concentration */
  glucose: number;
  /** Diastolic blood pressure (mm Hg) */
  bloodPressure: number;
  /** Triceps skin fold thickness (mm) */
  skinThickness: number;
  /** 2-Hour serum insulin (mu U/ml) */
  insulin: number;
  /** Body mass index */
  bmi: number;
  /** Diabetes pedigree function */
  diabetesPedigreeFunction: number;
  /** Age (years) */
  age: number;
}