# lib/api-zod/src/generated/types/heartInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multiple Disease Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface HeartInput {
  /** Age in years */
  age: number;
  /** Sex (1 = male, 0 = female) */
  sex: number;
  /** Chest pain type (0-3) */
  cp: number;
  /** Resting blood pressure (mm Hg) */
  trestbps: number;
  /** Serum cholestoral in mg/dl */
  chol: number;
  /** Fasting blood sugar > 120 mg/dl (1 = true, 0 = false) */
  fbs: number;
  /** Resting electrocardiographic results (0-2) */
  restecg: number;
  /** Maximum heart rate achieved */
  thalach: number;
  /** Exercise induced angina (1 = yes, 0 = no) */
  exang: number;
  /** ST depression induced by exercise */
  oldpeak: number;
  /** Slope of peak exercise ST segment (0-2) */
  slope: number;
  /** Number of major vessels colored by fluoroscopy (0-3) */
  ca: number;
  /** Thalassemia (0 = normal, 1 = fixed defect, 2 = reversable defect) */
  thal: number;
}