# lib/api-zod/src/generated/types/parkinsonsInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multiple Disease Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface ParkinsonsInput {
  /** MDVP Fo(Hz) - Average vocal fundamental frequency */
  mdvpFo: number;
  /** MDVP Fhi(Hz) - Maximum vocal fundamental frequency */
  mdvpFhi: number;
  /** MDVP Flo(Hz) - Minimum vocal fundamental frequency */
  mdvpFlo: number;
  /** MDVP Jitter(%) */
  mdvpJitter: number;
  /** MDVP Jitter(Abs) */
  mdvpJitterAbs: number;
  /** MDVP RAP */
  mdvpRap: number;
  /** MDVP PPQ */
  mdvpPpq: number;
  /** Jitter DDP */
  jitterDdp: number;
  /** MDVP Shimmer */
  mdvpShimmer: number;
  /** MDVP Shimmer(dB) */
  mdvpShimmerDb: number;
  /** Shimmer APQ3 */
  shimmerApq3: number;
  /** Shimmer APQ5 */
  shimmerApq5: number;
  /** MDVP APQ */
  mdvpApq: number;
  /** Shimmer DDA */
  shimmerDda: number;
  /** NHR - Noise-to-Harmonics Ratio */
  nhr: number;
  /** HNR - Harmonics-to-Noise Ratio */
  hnr: number;
  /** RPDE - Recurrence Period Density Entropy */
  rpde: number;
  /** DFA - Detrended Fluctuation Analysis */
  dfa: number;
  /** Spread1 - nonlinear measure */
  spread1: number;
  /** Spread2 - nonlinear measure */
  spread2: number;
  /** D2 - nonlinear dynamical complexity */
  d2: number;
  /** PPE - Pitch Period Entropy */
  ppe: number;
}