# lib/api-zod/src/generated/types/predictionResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Multiple Disease Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface PredictionResult {
  /** 0 = negative, 1 = positive */
  prediction: number;
  /** Human-readable result label */
  label: string;
  /** Confidence score between 0 and 1 */
  confidence?: number;
  /** Descriptive result message */
  message: string;
}